Update funcfile.py
commented out an if loop that was creating syntax errors. I'd like to fix it when I get around to debugging this code because it did improve CNV call concordance with bin count data in cells with highly altered genomes 

this is where failing but small CNV calls are merged with neighboring CNV calls IF both are on the same chromosome, have the same copy number (as each-other, ignore the small call), and pass FUnC (I need to look up if they were also required to be large). This is because those surrounding CNVs form a different baseline of CN estimate for that genomic region.